Concluindo atividade 8

+++ Atividade 8/biblioteca/app/Http/Controllers/AuthorController.php
@@ -2,11 +2,21 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Book;
 use App\Models\Author;
+use App\Models\Publisher;
+use App\Models\Category;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 
 class AuthorController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth');
+        // Certifique-se de que o middleware de autorização não está aplicando restrições não desejadas
+    }
+    
     // Função para exibir uma lista de autores
     public function index()
     {
@@ -24,12 +34,16 @@ class AuthorController extends Controller
     // Função para exibir o formulário de criação de um novo autor
     public function create()
     {
+        $this->authorize('create', Book::class);
+
         return view('authors.create');
     }
 
     // Função para armazenar um novo autor no banco de dados
     public function store(Request $request)
     {
+        $this->authorize('create', Book::class);
+
         $validatedData = $request->validate([
             'name' => 'required|string|max:255',
             'birth_date' => 'nullable|date',
@@ -43,6 +57,8 @@ class AuthorController extends Controller
     // Função para exibir o formulário de edição de um autor
     public function edit($id)
     {
+        $this->authorize('create', Book::class);
+
         $author = Author::findOrFail($id);
         return view('authors.edit', compact('author'));
     }
@@ -50,6 +66,8 @@ class AuthorController extends Controller
     // Função para atualizar um autor no banco de dados
     public function update(Request $request, $id)
     {
+        $this->authorize('create', Book::class);
+
         $validatedData = $request->validate([
             'name' => 'required|string|max:255',
             'birth_date' => 'nullable|date',
@@ -64,6 +82,8 @@ class AuthorController extends Controller
     // Função para excluir um autor do banco de dados
     public function destroy($id)
     {
+        $this->authorize('create', Book::class);
+
         $author = Author::findOrFail($id);
         $author->delete();
