+++ Atividade 8/biblioteca/app/Http/Controllers/CategoryController.php
@@ -2,11 +2,21 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Book;
+use App\Models\Author;
+use App\Models\Publisher;
 use App\Models\Category;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 
 class CategoryController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth');
+        // Certifique-se de que o middleware de autorização não está aplicando restrições não desejadas
+    }
+
     // Função para exibir uma lista de categorias
     public function index()
     {
@@ -24,12 +34,16 @@ class CategoryController extends Controller
     // Função para exibir o formulário de criação de uma nova categoria
     public function create()
     {
+        $this->authorize('create', Book::class);
+
         return view('categories.create');
     }
 
     // Função para armazenar uma nova categoria no banco de dados
     public function store(Request $request)
     {
+        $this->authorize('create', Book::class);
+
         $validatedData = $request->validate([
             'name' => 'required|string|max:255|unique:categories',
         ]);
@@ -42,6 +56,8 @@ class CategoryController extends Controller
     // Função para exibir o formulário de edição de uma categoria
     public function edit($id)
     {
+        $this->authorize('create', Book::class);
+
         $category = Category::findOrFail($id);
         return view('categories.edit', compact('category'));
     }
@@ -49,6 +65,8 @@ class CategoryController extends Controller
     // Função para atualizar uma categoria no banco de dados
     public function update(Request $request, $id)
     {
+        $this->authorize('create', Book::class);
+
         $validatedData = $request->validate([
             'name' => 'required|string|max:255|unique:categories,name,' . $id,
         ]);
@@ -62,6 +80,8 @@ class CategoryController extends Controller
     // Função para excluir uma categoria do banco de dados
     public function destroy($id)
     {
+        $this->authorize('create', Book::class);
+
         $category = Category::findOrFail($id);
         $category->books()->detach();
         $category->delete();
