+++ Atividade 8/biblioteca/app/Http/Controllers/PublisherController.php
@@ -2,11 +2,21 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Book;
+use App\Models\Author;
 use App\Models\Publisher;
+use App\Models\Category;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 
 class PublisherController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth');
+        // Certifique-se de que o middleware de autorização não está aplicando restrições não desejadas
+    }
+
     // Função para exibir uma lista de editoras
     public function index()
     {
@@ -24,12 +34,16 @@ class PublisherController extends Controller
     // Função para exibir o formulário de criação de uma nova editora
     public function create()
     {
+        $this->authorize('create', Book::class);
+
         return view('publishers.create');
     }
 
     // Função para armazenar uma nova editora no banco de dados
     public function store(Request $request)
     {
+        $this->authorize('create', Book::class);
+
         $validatedData = $request->validate([
             'name' => 'required|string|max:255',
             'address' => 'nullable|string|max:255',
@@ -43,6 +57,8 @@ class PublisherController extends Controller
     // Função para exibir o formulário de edição de uma editora
     public function edit($id)
     {
+        $this->authorize('create', Book::class);
+
         $publisher = Publisher::findOrFail($id);
         return view('publishers.edit', compact('publisher'));
     }
@@ -50,6 +66,8 @@ class PublisherController extends Controller
     // Função para atualizar uma editora no banco de dados
     public function update(Request $request, $id)
     {
+        $this->authorize('create', Book::class);
+
         $validatedData = $request->validate([
             'name' => 'required|string|max:255',
             'address' => 'nullable|string|max:255',
@@ -64,6 +82,8 @@ class PublisherController extends Controller
     // Função para excluir uma editora do banco de dados
     public function destroy($id)
     {
+        $this->authorize('create', Book::class);
+
         $publisher = Publisher::findOrFail($id);
         $publisher->delete();
